Создала модуль random, исправила ошибки
